+++ packages/svelte-undp-components/src/lib/components/CountrySelector.stories.ts
@@ -25,6 +25,20 @@ const meta = {
 			type: 'string',
 			description: 'Set GeoHub origin URL if you want to use it in other places',
 			defaultValue: ''
+		},
+		continents: {
+			description:
+				'An array of continent code returned from /api/continents API. If regions are set, continets will be ignored.',
+			defaultValue: []
+		},
+		regions: {
+			description: 'An array of region code returned from /api/regions API.',
+			defaultValue: []
+		},
+		showSelectedCountries: {
+			type: 'boolean',
+			description: 'If true, show selected countries in the top of tooltip.',
+			defaultValue: true
 		}
 	}
 } satisfies Meta<CountrySelector>;
@@ -52,5 +66,46 @@ export const Selected: Story = {
 	args: {
 		selected: ['KEN', 'RWA'],
 		geohubOrigin: 'https://dev.undpgeohub.org'
+	}
+};
+
+export const hideSelectedCountryChips: Story = {
+	args: {
+		selected: [],
+		showOnlyExists: true,
+		geohubOrigin: 'https://dev.undpgeohub.org',
+		showSelectedCountries: false
+	}
+};
+
+export const FilterByAfrica: Story = {
+	args: {
+		selected: [],
+		geohubOrigin: 'https://dev.undpgeohub.org',
+		continents: [2]
+	}
+};
+
+export const FilterByAfricaAndAsia: Story = {
+	args: {
+		selected: [],
+		geohubOrigin: 'https://dev.undpgeohub.org',
+		continents: [2, 142]
+	}
+};
+
+export const FilterByEastAsia: Story = {
+	args: {
+		selected: [],
+		geohubOrigin: 'https://dev.undpgeohub.org',
+		regions: [30]
+	}
+};
+
+export const FilterByCentralAndEastAsia: Story = {
+	args: {
+		selected: [],
+		geohubOrigin: 'https://dev.undpgeohub.org',
+		regions: [143, 30]
 	}
 };