+++ packages/svelte-undp-components/src/lib/components/SDGSelector.svelte
@@ -60,7 +60,7 @@
 </script>
 
 <div
-	class="button sdg-button {isFullWidth ? 'is-fullwidth' : ''} px-4"
+	class="button sdg-button {isFullWidth ? 'is-fullwidth' : ''} px-2"
 	aria-haspopup="true"
 	aria-controls="dropdown-menu"
 	use:tippy={{ content: tooltipContent }}
@@ -81,11 +81,7 @@
 	</span>
 </div>
 
-<div
-	bind:this={tooltipContent}
-	class="sdg-content tooltip"
-	style="width: {(buttonWidth > 500 ? 500 : buttonWidth) - 20}px;"
->
+<div bind:this={tooltipContent} class="sdg-content tooltip">
 	{#each sdgNumbers as number}
 		{@const isSelected = selected.includes(number)}
 		<div class="sdg-item py-2 pl-2 pr-4">
@@ -115,7 +111,6 @@
 	.sdg-button {
 		border: 1px solid black;
 		box-shadow: none;
-		min-width: 300px;
 
 		.toggle-icon {
 			-webkit-transition: all 0.3s ease;